Move SizeBinding builder into utils module
