+++ mvvmfx-utils/pom.xml
@@ -43,6 +43,13 @@
         </dependency>
 
         <!-- Testing Frameworks -->
+
+        <dependency>
+            <groupId>de.saxsys</groupId>
+            <artifactId>jfx-testrunner</artifactId>
+            <version>1.1</version>
+            <scope>test</scope>
+        </dependency>
         <dependency>
             <groupId>junit</groupId>
             <artifactId>junit</artifactId>
