+++ mvvmfx-utils/src/main/java/de/saxsys/mvvmfx/utils/sizebinding/internal/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Internal: Implementation of the {@link de.saxsys.mvvmfx.utils.sizebinding.SizeBindingsBuilder} step-builder.
  */
-package de.saxsys.mvvmfx.internal.utils.sizebinding;
+package de.saxsys.mvvmfx.utils.sizebinding.internal;
