+++ mvvmfx-utils/src/main/java/de/saxsys/mvvmfx/utils/sizebinding/package-info.java
